requirements.txt
discord.py
random.py